DI2018-70

+++ Dziennik/Controllers/DonosController.cs
@@ -4,6 +4,7 @@ using System.Data;
 using System.Data.Entity;
 using System.Linq;
 using System.Net;
+using System.Net.Mail;
 using System.Web;
 using System.Web.Mvc;
 using Dziennik.DAL;
@@ -18,12 +19,14 @@ namespace Dziennik.Controllers
         // GET: Donos
         public ActionResult Index()
         {
+            if ((string)Session["Status"] != "Uczen")
+                return RedirectToAction("Index", "Home");
             var donos = db.Donos.Include(d => d.Nauczyciel).Include(d => d.Uczen);
             return View(donos.ToList());
         }
 
         // GET: Donos/Details/5
-        public ActionResult Details(int? id)
+      /*  public ActionResult Details(int? id)
         {
             if (id == null)
             {
@@ -36,10 +39,12 @@ namespace Dziennik.Controllers
             }
             return View(donos);
         }
-
+        */
         // GET: Donos/Create
         public ActionResult Create()
         {
+            if ((string)Session["Status"] != "Uczen")
+                return RedirectToAction("Index", "Home");
             ViewBag.NauczycielID = new SelectList(db.Nauczyciele, "NauczycielID", "imie");
             //ViewBag.RodzicID = new SelectList(db.Rodzics, "ID", "imie");
             return View();
@@ -52,6 +57,8 @@ namespace Dziennik.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Create([Bind(Include = "ID,NauczycielID,RodzicID,pytanie,odpowiedz,data_pytania,data_odpowiedz")] Donos donos)
         {
+            if ((string)Session["Status"] != "Uczen")
+                return RedirectToAction("Index", "Home");
             if (ModelState.IsValid)
             {
                 db.Donos.Add(donos);
@@ -60,11 +67,35 @@ namespace Dziennik.Controllers
             }
 
             ViewBag.NauczycielID = new SelectList(db.Nauczyciele, "NauczycielID", "imie", donos.NauczycielID);
+            donos.data_pytania = DateTime.Now;
             var user = Session["UserID"];
             string ide = user.ToString();
             int id = Convert.ToInt32(ide);
             donos.UczenID = id;
             donos.Uczen = db.Uczniowie.Find(id);
+
+            var body = "<p>Email From: {0} ({1})</p><p>Message:</p><p>{2}</p>";
+            var message = new MailMessage();
+            message.To.Add(new MailAddress(donos.Nauczyciel.Email));  // replace with valid value 
+            message.From = new MailAddress("mojagracv@gmail.com");  // replace with valid value
+            message.Subject = "Wazna wiadomosc od "+donos.Uczen;
+            message.Body = string.Format(body, donos.Uczen.FullName,donos.data_pytania, donos.wiadomosc);
+            message.IsBodyHtml = true;
+
+            using (var smtp = new SmtpClient())
+            {
+                var credential = new NetworkCredential
+                {
+                    UserName = "mojagracv@gmail.com",  // replace with valid value
+                    Password = "civilization96"  // replace with valid value
+                };
+                smtp.Credentials = credential;
+                smtp.Host = "smtp.gmail.com";
+                smtp.Port = 587;
+                smtp.EnableSsl = true;
+                smtp.SendMailAsync(message);
+              
+            }
             return View(donos);
         }
 
@@ -105,6 +136,8 @@ namespace Dziennik.Controllers
         // GET: Donos/Delete/5
         public ActionResult Delete(int? id)
         {
+            if ((string)Session["Status"] != "Uczen")
+                return RedirectToAction("Index", "Home");
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -122,6 +155,8 @@ namespace Dziennik.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(int id)
         {
+            if ((string)Session["Status"] != "Uczen")
+                return RedirectToAction("Index", "Home");
             Donos donos = db.Donos.Find(id);
             db.Donos.Remove(donos);
             db.SaveChanges();
